+++ react/SummaryContext.tsx
@@ -1,11 +1,12 @@
-import React, { createContext, useContext, FunctionComponent } from 'react'
+import React, { createContext, useContext, PropsWithChildren } from 'react'
 
 import { SummaryProps } from './Summary'
 
 const SummaryContext = createContext<SummaryProps | undefined>(undefined)
 
-export const useSummary = () => {
+export function useSummary() {
   const context = useContext(SummaryContext)
+
   if (context === undefined) {
     throw new Error('useSummary must be used within a SummaryProvider')
   }
@@ -13,25 +14,27 @@ export const useSummary = () => {
   return context
 }
 
-const SummaryContextProvider: FunctionComponent<SummaryProps> = ({
+function SummaryContextProvider({
   coupon,
   insertCoupon,
   loading,
   totalizers,
   total,
   children,
-}) => (
-  <SummaryContext.Provider
-    value={{
-      coupon,
-      insertCoupon,
-      loading,
-      totalizers,
-      total,
-    }}
-  >
-    {children}
-  </SummaryContext.Provider>
-)
+}: PropsWithChildren<SummaryProps>) {
+  return (
+    <SummaryContext.Provider
+      value={{
+        coupon,
+        insertCoupon,
+        loading,
+        totalizers,
+        total,
+      }}
+    >
+      {children}
+    </SummaryContext.Provider>
+  )
+}
 
 export default SummaryContextProvider