+++ react/SummarySmall.tsx
@@ -1,37 +1,54 @@
-import React, { FunctionComponent } from 'react'
+import React, { PropsWithChildren } from 'react'
 import { FormattedMessage } from 'react-intl'
 import { useCssHandles } from 'vtex.css-handles'
 
+import { Totalizer } from './modules/types'
 import SummaryContextProvider from './SummaryContext'
 
 interface Props {
   totalizers: Totalizer[]
   total: number
-  totalizersToShow: string[]
+  totalizersToShow?: string[]
+  totalCalculation?: 'visibleTotalizers' | 'allTotalizers'
 }
 
 const CSS_HANDLES = ['summarySmallContent', 'summarySmallDisclaimer'] as const
 
-const SummarySmall: FunctionComponent<Props> = ({
+function SummarySmall({
   total,
   children,
   totalizers,
   totalizersToShow = ['Items'],
-}) => {
+  totalCalculation = 'visibleTotalizers',
+}: PropsWithChildren<Props>) {
   const handles = useCssHandles(CSS_HANDLES)
 
-  const filteredTotalizers = totalizers.filter(totalizer =>
+  const filteredTotalizers = totalizers.filter((totalizer) =>
     totalizersToShow.includes(totalizer.id)
   )
 
+  const totalToDisplay =
+    totalCalculation === 'visibleTotalizers'
+      ? filteredTotalizers.reduce((acc, totalizer) => {
+          return (acc += totalizer.value ?? 0)
+        }, 0)
+      : total
+
+  const shouldDisplayShippingDisclaimer = !totalizersToShow.includes('Shipping')
+
   return (
-    <SummaryContextProvider totalizers={filteredTotalizers} total={total}>
+    <SummaryContextProvider
+      totalizers={filteredTotalizers}
+      total={totalToDisplay}
+    >
       <div className={`${handles.summarySmallContent} c-on-base`}>
         {children}
       </div>
-      <span className={`${handles.summarySmallDisclaimer} t-small db mv4`}>
-        <FormattedMessage id="store/checkout-summary.disclaimer" />
-      </span>
+      {shouldDisplayShippingDisclaimer && (
+        <span className={`${handles.summarySmallDisclaimer} t-small db mv4`}>
+          <FormattedMessage id="store/checkout-summary.disclaimer" />
+        </span>
+      )}
     </SummaryContextProvider>
   )
 }