+++ react/modules/types.ts
@@ -1,4 +1,4 @@
-interface Totalizer {
+export interface Totalizer {
   id: string
   value: number | null
   name: string